requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
sqlparse==0.5.3
Werkzeug==3.1.3
